// src/contracts/generated/PresentationManifestV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CatalogId } from "./CatalogId";
import type { CatalogVersion } from "./CatalogVersion";
import type { LogoAssetsV1 } from "./LogoAssetsV1";
import type { MessageKey } from "./MessageKey";
import type { PresentationLayoutV1 } from "./PresentationLayoutV1";

export type PresentationManifestV1 = { schemaVersion: number, presentationId: CatalogId, presentationVersion: CatalogVersion, productNameMessageKey: MessageKey, logoAssets: LogoAssetsV1, artworkAssetId: CatalogId | null, themePairingId: CatalogId, layout: PresentationLayoutV1, };